int a long

+++ src/gestorAplicacion/clasesPrincipales/Conductor.java
@@ -8,13 +8,13 @@ public class Conductor implements Serializable {
 	private static final long serialVersionUID = 2L;
 	
 	private String nombre;
-	private int cedula;
+	private long cedula;
 	private long celular;
 	private int sueldo;
 	private boolean disponible;
 	private static ArrayList<Conductor> conductores = new ArrayList<>();
 	
-	public Conductor(String nombre, int cedula, long celular, int sueldo) {
+	public Conductor(String nombre, long cedula, long celular, int sueldo) {
 		this.nombre = nombre;
 		this.cedula = cedula;
 		this.celular = celular;
@@ -32,7 +32,7 @@ public class Conductor implements Serializable {
 		this.nombre = nombre;
 	}
 	
-	public int getCedula() {
+	public long getCedula() {
 		return cedula;
 	}
